+++ src/ecs/components/velocity.h
@@ -5,6 +5,6 @@
 // TODO: maybe positions, velocities, directions
 // should have been represented with same strucutre Vec2 {X, Y} ?
 typedef struct Velocity {
-    f32 X;
-    f32 Y;
+    f64 X;
+    f64 Y;
 } Velocity;
